refactor: resolve circular dependency by moving SourceFile and Span
Move SourceFile and Span types from tribute-ast to tribute-database
to eliminate circular dependency between packages. This creates a
clean unidirectional dependency flow and allows consistent use of
TributeDatabaseImpl throughout the codebase.
