+++ crates/tribute-ast/src/lib.rs
@@ -3,9 +3,7 @@
 //! This crate provides the core data structures and parsing logic for the Tribute language.
 
 pub mod ast;
-pub mod database;
 pub mod parser;
 
 pub use ast::*;
-pub use database::*;
 pub use parser::*;
