+++ src/lib.rs
@@ -4,8 +4,8 @@ pub mod eval;
 use std::path::Path;
 
 pub use crate::eval::{Environment, Value, eval_hir_expr, eval_hir_program};
-pub use tribute_ast::{Item, Program, SourceFile, TributeParser, ast, parse_source_file};
-pub use tribute_database::{Diagnostic, DiagnosticSeverity, TributeDatabaseImpl};
+pub use tribute_ast::{Item, Program, TributeParser, ast, parse_source_file};
+pub use tribute_database::{Diagnostic, DiagnosticSeverity, SourceFile, TributeDatabaseImpl};
 pub use tribute_hir::{compile_to_hir, lower_source_to_hir};
 
 /// Parse a Tribute source string and return the program along with diagnostics
